src/requirements.txt
# ---------------------------------------------------------------------
# Modeling
sqlmodel
# ---------------------------------------------------------------------
# Passwords and other user verification libraries
passlib
argon2-cffi
pyotp
# ---------------------------------------------------------------------